made img responsive; added success failure pages
